addressing further changes
